+++ LayoutTests/platform/glib/TestExpectations
@@ -698,6 +698,8 @@ accessibility/aria-table-selection-support.html [ Skip ]
 accessibility/dynamic-aria-describedby-subtree.html [ Skip ]
 accessibility/dynamic-aria-describedby-text.html [ Skip ]
 
+accessibility/button-in-deep-dom.html [ Skip ]
+
 # Timing out since added in https://bugs.webkit.org/show_bug.cgi?id=263715
 accessibility/focus-inside-modal.html [ Skip ]
 # Timing out since added in https://bugs.webkit.org/show_bug.cgi?id=249283.
