+++ Source/WebCore/dom/ComposedTreeIterator.cpp
@@ -26,195 +26,9 @@
 #include "config.h"
 #include "ComposedTreeIterator.h"
 
-#include "ElementInlines.h"
-#include "ElementRareData.h"
-#include "HTMLSlotElement.h"
 #include <wtf/text/TextStream.h>
 
 namespace WebCore {
-
-ComposedTreeIterator::Context::Context()
-{
-}
-
-ComposedTreeIterator::Context::Context(ContainerNode& root, FirstChildTag)
-    : iterator(root, ElementAndTextDescendantIterator::FirstChild)
-{
-}
-
-ComposedTreeIterator::Context::Context(ContainerNode& root, Node& node)
-    : iterator(root, &node)
-{
-}
-
-ComposedTreeIterator::Context::Context(ContainerNode& root, Node& node, SlottedTag)
-    : iterator(root, &node)
-    , end(iterator)
-{
-    end.traverseNextSkippingChildren();
-}
-
-ComposedTreeIterator::ComposedTreeIterator(ContainerNode& root, FirstChildTag)
-    : m_rootIsInShadowTree(root.isInShadowTree())
-{
-    ASSERT(!is<ShadowRoot>(root));
-
-    if (auto* slot = dynamicDowncast<HTMLSlotElement>(root)) {
-        if (auto* assignedNodes = slot->assignedNodes()) {
-            initializeContextStack(root, *assignedNodes->at(0));
-            return;
-        }
-    }
-    if (auto* shadowRoot = root.shadowRoot()) {
-        ElementAndTextDescendantIterator firstChild(*shadowRoot, ElementAndTextDescendantIterator::FirstChild);
-        initializeContextStack(root, firstChild ? *firstChild : root);
-        return;
-    }
-
-    m_contextStack.append(Context(root, FirstChild));
-}
-
-ComposedTreeIterator::ComposedTreeIterator(ContainerNode& root, Node& current)
-    : m_rootIsInShadowTree(root.isInShadowTree())
-{
-    ASSERT(!is<ShadowRoot>(root));
-    ASSERT(!is<ShadowRoot>(current));
-
-    bool mayNeedShadowStack = root.shadowRoot() || (&current != &root && current.parentNode() != &root);
-    if (mayNeedShadowStack)
-        initializeContextStack(root, current);
-    else
-        m_contextStack.append(Context(root, current));
-}
-
-void ComposedTreeIterator::initializeContextStack(ContainerNode& root, Node& current)
-{
-    // This code sets up the iterator for arbitrary node/root pair. It is not needed in common cases
-    // or completes fast because node and root are close (like in composedTreeChildren(*parent).at(node) case).
-    auto* node = &current;
-    auto* contextCurrent = node;
-    size_t currentSlotNodeIndex = notFound;
-    while (node != &root) {
-        auto* parent = node->parentNode();
-        if (!parent) {
-            *this = { };
-            return;
-        }
-        if (auto* shadowRoot = dynamicDowncast<ShadowRoot>(*parent)) {
-            m_contextStack.append(Context(*shadowRoot, *contextCurrent));
-            m_contextStack.last().slotNodeIndex = currentSlotNodeIndex;
-
-            node = shadowRoot->host();
-            contextCurrent = node;
-            currentSlotNodeIndex = notFound;
-            continue;
-        }
-        if (auto* shadowRoot = parent->shadowRoot()) {
-            m_contextStack.append(Context(*parent, *contextCurrent, Context::Slotted));
-            m_contextStack.last().slotNodeIndex = currentSlotNodeIndex;
-
-            auto* assignedSlot = shadowRoot->findAssignedSlot(*node);
-            if (assignedSlot) {
-                currentSlotNodeIndex = assignedSlot->assignedNodes()->find(node);
-                ASSERT(currentSlotNodeIndex != notFound);
-                node = assignedSlot;
-                contextCurrent = assignedSlot;
-                continue;
-            }
-            // The node is not part of the composed tree.
-            *this = { };
-            return;
-        }
-        node = parent;
-    }
-    m_contextStack.append(Context(root, *contextCurrent));
-    m_contextStack.last().slotNodeIndex = currentSlotNodeIndex;
-
-    m_contextStack.reverse();
-}
-
-void ComposedTreeIterator::dropAssertions()
-{
-    for (auto& context : m_contextStack)
-        context.iterator.dropAssertions();
-    m_didDropAssertions = true;
-}
-
-void ComposedTreeIterator::traverseShadowRoot(ShadowRoot& shadowRoot)
-{
-    Context shadowContext(shadowRoot, FirstChild);
-    if (!shadowContext.iterator) {
-        // Empty shadow root.
-        traverseNextSkippingChildren();
-        return;
-    }
-
-    if (m_didDropAssertions)
-        shadowContext.iterator.dropAssertions();
-
-    m_contextStack.append(WTFMove(shadowContext));
-}
-
-void ComposedTreeIterator::traverseNextInShadowTree()
-{
-    ASSERT(m_contextStack.size() > 1 || m_rootIsInShadowTree);
-
-    if (auto* slot = dynamicDowncast<HTMLSlotElement>(current())) {
-        if (auto* assignedNodes = slot->assignedNodes()) {
-            context().slotNodeIndex = 0;
-            auto* assignedNode = assignedNodes->at(0).get();
-            ASSERT(assignedNode);
-            ASSERT(assignedNode->parentElement());
-            m_contextStack.append(Context(*assignedNode->parentElement(), *assignedNode, Context::Slotted));
-            return;
-        }
-    }
-
-    context().iterator.traverseNext();
-
-    if (context().iterator == context().end)
-        traverseNextLeavingContext();
-}
-
-void ComposedTreeIterator::traverseNextLeavingContext()
-{
-    while (context().iterator == context().end && m_contextStack.size() > 1) {
-        m_contextStack.removeLast();
-        if (auto* slot = dynamicDowncast<HTMLSlotElement>(current()); slot && advanceInSlot(1, *slot))
-            return;
-        if (context().iterator == context().end)
-            return;
-        context().iterator.traverseNextSkippingChildren();
-    }
-}
-
-bool ComposedTreeIterator::advanceInSlot(int direction, const HTMLSlotElement& slot)
-{
-    ASSERT(context().slotNodeIndex != notFound);
-
-    auto& assignedNodes = *slot.assignedNodes();
-    // It is fine to underflow this.
-    context().slotNodeIndex += direction;
-    if (context().slotNodeIndex >= assignedNodes.size())
-        return false;
-
-    auto* slotNode = assignedNodes.at(context().slotNodeIndex).get();
-    ASSERT(slotNode);
-    ASSERT(slotNode->parentElement());
-    m_contextStack.append(Context(*slotNode->parentElement(), *slotNode, Context::Slotted));
-    return true;
-}
-
-void ComposedTreeIterator::traverseSiblingInSlot(int direction)
-{
-    ASSERT(m_contextStack.size() > 1);
-    ASSERT(current().parentNode()->shadowRoot());
-
-    m_contextStack.removeLast();
-
-    if (!advanceInSlot(direction, downcast<HTMLSlotElement>(current())))
-        *this = { };
-}
 
 String composedTreeAsText(ContainerNode& root, ComposedTreeAsTextMode mode)
 {
